+++ src/qibolab/backends.py
@@ -6,10 +6,10 @@ from qibo.backends import NumpyBackend
 from qibo.config import raise_error
 from qibo.states import CircuitResult
 
-from qibolab import ExecutionParameters
 from qibolab import __version__ as qibolab_version
 from qibolab import create_platform
 from qibolab.compilers import Compiler
+from qibolab.execution_parameters import ExecutionParameters
 from qibolab.platform import Platform
 from qibolab.transpilers import Pipeline
