+++ src/qibolab/platform.py
@@ -368,7 +368,7 @@ class Platform:
                 from qibolab.dummy import create_dummy
                 from qibolab.sweeper import Sweeper, Parameter
                 from qibolab.pulses import PulseSequence
-                from qibolab import ExecutionParameters
+                from qibolab.execution_parameters import ExecutionParameters
 
 
                 platform = create_dummy()
