+++ tests/test_instruments_qm.py
@@ -2,7 +2,10 @@ import numpy as np
 import pytest
 from qm import qua
 
-from qibolab import AcquisitionType, ExecutionParameters, create_platform
+from qibolab import create_platform
+from qibolab.execution_parameters import AcquisitionType, ExecutionParameters
+
+# Are you testing Singleshot or Averaging ?
 from qibolab.instruments.qm import QMOPX, Acquisition, QMPort, QMPulse, Sequence
 from qibolab.pulses import FluxPulse, Pulse, ReadoutPulse, Rectangular
